+++ src/Telegram.Bot/Exceptions/ApiExceptionParser.cs
@@ -13,6 +13,7 @@ namespace Telegram.Bot.Exceptions
             new BadRequestExceptionInfo<InvalidUserIdException>("USER_ID_INVALID"),
             new BadRequestExceptionInfo<InvalidStickerSetNameException>("sticker set name invalid"),
             new BadRequestExceptionInfo<InvalidStickerEmojisException>("invalid sticker emojis"),
+            new BadRequestExceptionInfo<InvalidStickerDimensionsException>("STICKER_PNG_DIMENSIONS"),
             new BadRequestExceptionInfo<ContactRequestException>("phone number can be requested in a private chats only"),
 
             new ForbiddenExceptionInfo<ChatNotInitiatedException>("bot can't initiate conversation with a user"),